refactor: add code store for sharing code utils between editor & renderer
